backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
openai==2.32.0
python-multipart==0.0.19
pydantic==2.10.3
python-dotenv==1.0.1
